readme and tf (AGIC) changes
Update README and main.tf for AGIC integration

Clarified that the Application Gateway Ingress Controller (AGIC) is automatically enabled as an AKS add-on, removing the need for manual installation. Expanded the example `terraform.tfvars` with additional variables and updated deployment instructions accordingly. Added a new block in main.tf for the AGIC Addon Profile to configure integration with the Application Gateway.